Reorganization of project
Organize project by:
Project Version->Vim->TestedOn
